Add refactor for listingskills, listingupdates, seeder, and advancedResults, pending fix hashtags bug
